Refactor color schemes changing per schema
